#115465: java.lang.AssertionError: Cache C:\Documents and Settings\Administrador\.netbeans\dev\var\cache\catalogcache\URL_efh exists.

+++ autoupdate.services/src/org/netbeans/modules/autoupdate/updateprovider/AutoupdateCatalogCache.java
@@ -192,9 +192,17 @@ public class AutoupdateCatalogCache {
                 if (os != null) os.flush ();
                 if (os != null) os.close ();
                 synchronized (this) {
-                    cache.delete ();
+                    if (cache.exists () && ! cache.delete ()) {
+                        err.log (Level.INFO, "Cannot delete cache " + cache);
+                        try {
+                            Thread.sleep(200);
+                        } catch (InterruptedException ie) {
+                            assert false : ie;
+                        }
+                        cache.delete();
+                    }
                     if (! temp.renameTo (cache)) {
-                        throw new IOException ("Cannot move temp " + temp + " to cache " + cache);
+                        err.log (Level.INFO, "Cannot rename temp " + temp + " to cache " + cache);
                     }
                 }                    
             } catch (IOException ioe) {
